Refactor of Cargo.toml Division of main.rs file and run functions into 2 files (run.rs)
