+++ owlyshield_predict/Cargo.toml
@@ -4,19 +4,11 @@ version = "1.0.1"
 edition = "2021"
 license-file = "LICENSE.txt"
 
-#[package]
-#name = "redbpf_test"
-#version = "0.1.0"
-#edition = "2018"
-
-
 [dependencies]
 moonfire-tflite = { path = "moonfire-tflite" }
 slc-paths = { path = "slc-paths" }
 win-pe-inspection = { path = "win-pe-inspection" }
 wchar = "0.11.0"
-widestring = { version = "1.0.2", optional = true }
-sysinfo = "0.20.3"
 registry = "1.2.0"
 strum = "0.24.1"
 strum_macros = "0.24.2"
@@ -33,25 +25,18 @@ hostname = "0.3.1"
 curl = "0.4.43"
 jsonrpc-http-server = "18.0.0"
 configparser = "3.0.2"
-
-winlog = { version = "0.2.6", optional = true }
-windows-service = { version = "0.5.0", optional = true }
-
-redbpf = { version = "2.3.0", features = ["load"], optional = true }
-redbpf-probes = { version = "2.3.0", optional = true }
-tokio = { version = "1.0", features = ["rt", "signal", "time", "io-util", "net", "sync"], optional = true }
-tracing-subscriber = { version = "0.2", optional = true }
-tracing = { version = "0.1", optional = true }
-futures = { version = "0.3", optional = true }
-psutil = { version = "3.2.2", optional = true }
-#lru = { version = "0.8.0", optional = true }
-
-probes = { path = "./probes", optional = true }
 rumqtt = "0.31.0"
-lru = "0.8.0"
+lru = "0.9.0"
+sysinfo = "0.27.7"
 
-[dependencies.windows]
-version = "0.39.0"
+#Windows
+[target.'cfg(target_os = "windows")'.dependencies]
+winlog = { version = "0.2.6", features=["env_logger"] }
+windows-service = "0.5.0"
+widestring = "1.0.2"
+
+[target.'cfg(target_os = "windows")'.dependencies.windows]
+version = "0.44.0"
 features = [
     "Win32_Foundation",
     "Win32_Security",
@@ -62,18 +47,26 @@ features = [
     "Win32_System_Threading",
     "Win32_System_RemoteDesktop",
 ]
-optional = true
+
+# Linux
+[target.'cfg(target_os = "linux")'.dependencies]
+redbpf = { version = "2.3.0", features = ["load"] }
+redbpf-probes = "2.3.0"
+tokio = { version = "1.0", features = ["rt", "signal", "time", "io-util", "net", "sync"] }
+tracing-subscriber = "0.2"
+tracing = "0.1"
+futures = "0.3"
+psutil = "3.2.2"
+probes = { path = "./probes" }
 
 [profile.release]
 debug = true
 
 [features]
 default = []
-service = ["win"]
+service = []
 malware = []
 record = []
 replay = []
 jsonrpc = []
-linux = ["redbpf", "redbpf-probes", "tokio", "tracing-subscriber", "tracing", "futures", "psutil", "probes"]
-win = ["windows", "winlog", "windows-service", "widestring"]
 mqtt = []
