+++ owlyshield_predict/probes/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "probes"
 version = "0.1.0"
-edition = '2018'
+edition = '2021'
 
 [dependencies]
 cty = "0.2"
