+++ owlyshield_predict/src/connectors/community.rs
@@ -26,7 +26,7 @@ const CONF_LOCATION: &str = r"SOFTWARE\Owlyshield\Telemetry";
 const CONF_BLOC: &str = "";
 
 #[cfg(target_os = "linux")]
-const CONF_LOCATION: &str = "/etc/conf/owlyshield.conf";
+const CONF_LOCATION: &str = "/etc/owlyshield/owlyshield.conf";
 #[cfg(target_os = "linux")]
 const CONF_BLOC: &str = "telemetry";
