+++ owlyshield_predict/src/windows/driver_com.rs
@@ -20,7 +20,7 @@ use windows::Win32::Storage::InstallableFileSystems::{
 use self::shared_def::ReplyIrp;
 use self::IrpMajorOp::{IrpCreate, IrpNone, IrpRead, IrpSetInfo, IrpWrite};
 
-type BufPath = [wchar_t; 520];
+pub type BufPath = [wchar_t; 520];
 /// The usermode app (this app) can send several messages types to the driver. See [DriverComMessageType]
 /// for details.
 /// Depending on the message type, the *pid*, *gid* and *path* fields can be optional.
@@ -142,7 +142,7 @@ impl Driver {
 
         let mut get_irp_msg: DriverComMessage = DriverComMessage {
             r#type: DriverComMessageType::MessageSetPid as c_ulong,
-            pid: get_current_pid().unwrap() as c_ulong,
+            pid: usize::from(get_current_pid().unwrap()) as c_ulong,
             gid: 140713315094899,
             path: buf, //wch!("\0"),
         };
@@ -152,7 +152,7 @@ impl Driver {
                 self.handle,
                 ptr::addr_of_mut!(get_irp_msg) as *mut c_void,
                 mem::size_of::<DriverComMessage>() as c_ulong,
-                ptr::null_mut(),
+                Some(ptr::null_mut()),
                 0,
                 &mut tmp as *mut u32,
             )
@@ -171,9 +171,9 @@ impl Driver {
             handle = FilterConnectCommunicationPort(
                 PCWSTR(com_port_name),
                 0,
-                ptr::null(),
+                Some(ptr::null()),
                 0,
-                ptr::null_mut(),
+                Some(ptr::null_mut()),
             )?
         }
         let res = Driver { handle };
@@ -196,7 +196,7 @@ impl Driver {
                 self.handle,
                 ptr::addr_of_mut!(get_irp_msg) as *mut c_void,
                 mem::size_of::<DriverComMessage>() as c_ulong,
-                vecnew.as_ptr() as *mut c_void,
+                Some(vecnew.as_ptr() as *mut c_void),
                 65536,
                 ptr::addr_of_mut!(tmp) as *mut u32,
             )
@@ -229,7 +229,7 @@ impl Driver {
                 self.handle,
                 ptr::addr_of_mut!(killmsg) as *mut c_void,
                 mem::size_of::<DriverComMessage>() as c_ulong,
-                ptr::addr_of_mut!(res) as *mut c_void,
+                Some(ptr::addr_of_mut!(res) as *mut c_void),
                 4,
                 ptr::addr_of_mut!(res_size) as *mut u32,
             )?;
@@ -255,7 +255,7 @@ impl Driver {
     ) -> DriverComMessage {
         DriverComMessage {
             r#type: commsgtype as c_ulong, // MessageSetPid
-            pid: pid as c_ulong,
+            pid: usize::from(pid) as c_ulong,
             gid,
             path: Driver::string_to_commessage_buffer(path),
         }
